Add base of program to reorder pixel with an injunction table

+++ ctapipe_io_mchdf5/tools/mchdf5_injtab_sort.py
@@ -123,7 +123,7 @@ def main():
 	inputFileName = args.input
 	outputFileName = args.output
 	
-	injunctionTable = np.fromfile(args.injtab)
+	injunctionTable = np.fromfile(args.injtab, dtype=np.uint16)
 	
 	isStoreSlicePixel = True
 	if args.pixelslice != None:
